move deletevideo and deleteuser in a service
